adding change event listener
to change the game for 5 to 10 points is better to use the change event.

"this" refers to the specific object that i am listening for the event on (winningscoreSelect)

winningScore when i achange i get a string but i want a number. to change a number a refer to "PARSE INT"