adding reset function
adding a own fucntion for reset.
i move the code from resetbutton to reset funtion. then just call the reset function in the reset button so when the user click on the reset button, reset function is called.

as well calling the reset function when the user change the wiining score